[Data] Moving Things into Packages
