+++ app/src/main/java/me/rikinmarfatia/hail/MainActivity.kt
@@ -3,7 +3,7 @@ package me.rikinmarfatia.hail
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import me.rikinmarfatia.hail.features.feed.WeatherFeed
+import me.rikinmarfatia.hail.features.weather.presentation.WeatherFeed
 import me.rikinmarfatia.hail.ui.theme.HailTheme
 
 class MainActivity : ComponentActivity() {
