+++ app/src/main/java/me/rikinmarfatia/hail/features/weather/data/WeatherRepository.kt
@@ -1,24 +1,7 @@
-package me.rikinmarfatia.hail
+package me.rikinmarfatia.hail.features.weather.data
 
-import com.squareup.moshi.Json
-import com.squareup.moshi.JsonClass
 import retrofit2.Retrofit
 import retrofit2.converter.moshi.MoshiConverterFactory
-
-@JsonClass(generateAdapter = true)
-data class WeatherFeed(
-    @Json(name = "consolidated_weather") val consolidatedWeather: List<Weather>,
-    @Json(name = "title") val title: String
-)
-
-@JsonClass(generateAdapter = true)
-data class Weather(
-    @Json(name = "min_temp") val minTemp: Double,
-    @Json(name = "max_temp") val maxTemp: Double,
-    @Json(name = "the_temp") val theTemp: Double,
-    @Json(name = "applicable_date") val applicableDate: String,
-    @Json(name = "weather_state_abbr") val weatherStateAbbr: String
-)
 
 interface WeatherRepository {
     suspend fun getWeather(): WeatherFeed
